ImagesTable: Rename `ImagesTable.tsx` to `index.tsx`
This renames the original table file to index.
